+++ app/src/main/java/com/picpay/desafio/android/datasource/repository/UserRepository.kt
@@ -1,11 +1,9 @@
-package com.picpay.desafio.android.data.repository
+package com.picpay.desafio.android.datasource.repository
 
-import com.picpay.desafio.android.data.mappers.toDomain
-import com.picpay.desafio.android.data.remote.data_source.UserRDS
+import com.picpay.desafio.android.datasource.mappers.toDomain
+import com.picpay.desafio.android.datasource.remote.UserRDS
 import com.picpay.desafio.android.domain.model.User
 import com.picpay.desafio.android.domain.repository.UserDataRepository
-import retrofit2.HttpException
-import java.net.SocketException
 
 class UserRepository(private val userRDS: UserRDS) :UserDataRepository{
     override suspend fun getUsers(): List<User> {
