+++ app/src/main/java/com/picpay/desafio/android/ui/UserListDiffCallback.kt
@@ -1,7 +1,6 @@
 package com.picpay.desafio.android.ui
 
 import androidx.recyclerview.widget.DiffUtil
-import com.picpay.desafio.android.data.remote.model.UserRM
 import com.picpay.desafio.android.domain.model.User
 
 class UserListDiffCallback(
