+++ app/src/main/java/com/picpay/desafio/android/ui/UserListItemViewHolder.kt
@@ -3,7 +3,6 @@ package com.picpay.desafio.android.ui
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import com.picpay.desafio.android.R
-import com.picpay.desafio.android.data.remote.model.UserRM
 import com.picpay.desafio.android.domain.model.User
 import com.squareup.picasso.Callback
 import com.squareup.picasso.Picasso
