+++ app/src/test/java/com/picpay/desafio/android/ExampleServiceTest.kt
@@ -2,8 +2,7 @@ package com.picpay.desafio.android
 
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
-import com.picpay.desafio.android.data.remote.data_source.UserRDS
-import com.picpay.desafio.android.data.remote.model.UserRM
+import com.picpay.desafio.android.datasource.remote.UserRDS
 import com.picpay.desafio.android.domain.model.User
 import junit.framework.Assert.assertEquals
 import org.junit.Test
